refactor(chart): split plugins.ts into per-plugin files (#57)
The 600-line chart/plugins.ts grew organically through v0.6 → v1.x as
each new plugin was appended. SonarCloud flagged the
dailyTickLabelsPlugin afterDraw at cognitive complexity 46 (limit 15)
— the worst function in the codebase.

Restructure: one plugin per file under chart/plugins/, keep the old
chart/plugins.ts as a barrel re-export so existing import sites
work unchanged. The barrel:

  chart/plugins.ts                     (re-exports — was the monolith)
  chart/plugins/_shared.ts             (ChartLike, ChartPlugin, …)
  chart/plugins/separator.ts           (createSeparatorPlugin)
  chart/plugins/daily-tick-labels.ts   (createDailyTickLabelsPlugin)
  chart/plugins/precip-label.ts        (createPrecipLabelPlugin)
  chart/plugins/sunshine-label.ts      (createSunshineLabelPlugin)

Plus extracts inside daily-tick-labels.ts that drop the afterDraw
hook from cog 46 to ~7:

  drawHourlyTimeLabels(chart, ctx)         — 'today' / 'hourly' branch
  drawDailyDateWeekdayLabels(chart, ctx)   — 'daily' branch
  findLeftmostVisibleTick(chart, scrollLeft)  — viewport scan
  formatSunshineHours(value)               — pretty-print Xh

The afterDraw hook is now a 4-line dispatch on `forecast.type`. Each
branch helper owns its own loop, save/restore, and column geometry.

Test setup: vitest.config.js's coverage.include now lists each new
plugin file so per-file thresholds apply individually (rather than
hiding under the previous file aggregate).

No behaviour change. Bundle unchanged (rollup tree-shakes the barrel
back into one chunk). dist diff is byte-identical.

Closes part of #57 (the cognitive-complexity headline). The remaining
nested-ternary / MINOR backlog continues in a follow-up commit.
