+++ vitest.config.js
@@ -20,7 +20,14 @@ export default {
         'src/forecast-utils.ts',
         'src/sunshine-source.ts',
         'src/openmeteo-source.ts',
+        // chart/plugins.ts is now a barrel re-export (#57); the
+        // actual plugin code lives in chart/plugins/*.ts.
         'src/chart/plugins.ts',
+        'src/chart/plugins/_shared.ts',
+        'src/chart/plugins/separator.ts',
+        'src/chart/plugins/daily-tick-labels.ts',
+        'src/chart/plugins/precip-label.ts',
+        'src/chart/plugins/sunshine-label.ts',
         'src/scroll-ux.ts',
         'src/action-handler.ts',
         'src/teardown-registry.ts',
